requirements.txt
django==5.2.1
djangorestframework==3.16.0
pillow==11.2.1
djangorestframework-simplejwt==5.5.0